Add convenience methods
Add convenience method to get all the main categories of fields from a
record.

It's much simpler (and a little faster) to do string comparisons on the tags, so
rather than getting the matching node from the XML and instantiating the
DataField or ControlField, we now just instantiate all the fields when the
record is initialized and select the matching records from the array. Add
private tag_range method to accomplish this and update some methods to use
tag_range.

Update tests for the new methods.

Rewrote the README to be simpler.