Remove getting fields by array and integer tags
The ability to call #field with and array of tags was really a side effect of
how the `#title` and `#mainEntry` methods were first written. This was
complicated to document, especially when subfields were involved, and in the end
is was simpler to use tag_range for the convenience method. Remove all code that
takes and array of tags to return fields, and tests for this behavior.

Also, because of the likely gotcha that someone might use an octal number as a
numeric tag (like `010`, which Ruby will see as octal for 8), only accept
strings as tags. Update tests.

Update `#mainEntry` to use `#tag_range`.